Move encodingutils to pkg
