+++ pkg/encodingutils/base64utils/base64utils_test.go
@@ -1,0 +1,21 @@
+package base64utils_test
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	"github.com/asciich/asciichgolangpublic/pkg/encodingutils/base64utils"
+)
+
+func TestBase64_encodeAndDecode(t *testing.T) {
+	t.Run("hello world", func(t *testing.T) {
+		encoded, err := base64utils.EncodeStringAsString("hello world")
+		require.NoError(t, err)
+
+		decoded, err := base64utils.DecodeStringAsString(encoded)
+		require.NoError(t, err)
+
+		require.EqualValues(t, "hello world", decoded)
+	},
+	)
+}
