+++ pkg/encodingutils/utf16utils/UTF16Utils.go
