+++ pkg/encodingutils/utf16utils/UTF16Utils_test.go
@@ -1,9 +1,10 @@
-package utf16utils
+package utf16utils_test
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+	"github.com/asciich/asciichgolangpublic/pkg/encodingutils/utf16utils"
 	"github.com/asciich/asciichgolangpublic/testutils"
 )
 
@@ -28,12 +29,9 @@ func TestUTF16DecodeAsString(t *testing.T) {
 		t.Run(
 			testutils.MustFormatAsTestname(tt),
 			func(t *testing.T) {
-				require := require.New(t)
-
-				require.EqualValues(
-					tt.expectedOutput,
-					MustDecodeAsString(tt.inputUtf16),
-				)
+				decoded, err := utf16utils.DecodeAsString(tt.inputUtf16)
+				require.NoError(t, err)
+				require.EqualValues(t, tt.expectedOutput, decoded)
 			},
 		)
 	}
